Added events and changed the name of patches +Added "AnnounceDecontaminationEvent". +Added "AnnounceScpTerminationEvent". +Added "AnnounceNtfEntranceEvent". +Added "SpawnRagdollEvent". +Changed the name of a lot of patches to bring more consistency: If a patch fixes a game bug, it will get a "Fix" suffix (like MuteHandlerFix). If a patch adds one or more events to EXILED, it will be called with the name of the event or with a generic name if there are more than one event in the same class (like ItemChangedEvent or GeneratorEvents). Otherwise, it will get the "Patch" suffix (like AntiFlyPatch).
