+++ EXILED_Events/EXILED_Events.csproj
@@ -174,53 +174,57 @@
     <Compile Include="Events\ServerEvents.cs" />
     <Compile Include="ExiledVersion.cs" />
     <Compile Include="LevenshteinDistance.cs" />
-    <Compile Include="Patches\079TriggerTeslaEvent.cs" />
-    <Compile Include="Patches\914MachineOverride.cs" />
-    <Compile Include="Patches\AntiFlyOverride.cs" />
-    <Compile Include="Patches\BanHandlerOverride.cs" />
-    <Compile Include="Patches\BanUserOverride.cs" />
-    <Compile Include="Patches\CallCmdMakePortalOverride.cs" />
-    <Compile Include="Patches\CheaterReportOverride.cs" />
+    <Compile Include="Patches\AnnounceDecontaminationEvent.cs" />
+    <Compile Include="Patches\AnnounceNtfEntranceEvent.cs" />
+    <Compile Include="Patches\AnnounceScpTerminationEvent.cs" />
+    <Compile Include="Patches\Scp079TriggerTeslaEvent.cs" />
+    <Compile Include="Patches\SCP914UpgradeEvent.cs" />
+    <Compile Include="Patches\AntiFlyPatch.cs" />
+    <Compile Include="Patches\PlayerBannedEvent.cs" />
+    <Compile Include="Patches\PlayerBanEvent.cs" />
+    <Compile Include="Patches\Scp106CreatedPortalEvent.cs" />
+    <Compile Include="Patches\CheaterReportEvent.cs" />
     <Compile Include="Patches\CheckEscapeEvent.cs" />
     <Compile Include="Patches\CheckRoundEndEvent.cs" />
-    <Compile Include="Patches\ElevatorInteractionEvent.cs" />
+    <Compile Include="Patches\ElevatorInteractEvent.cs" />
     <Compile Include="Patches\GrenadeExplosionPatch.cs" />
-    <Compile Include="Patches\LockerEventOverride.cs" />
+    <Compile Include="Patches\LockerInteractEvent.cs" />
     <Compile Include="Patches\MuteHandlerFix.cs" />
     <Compile Include="Patches\Scp079Events.cs" />
-    <Compile Include="Patches\CallCmdSetUnicOverride.cs" />
+    <Compile Include="Patches\ItemChangedEvent.cs" />
+    <Compile Include="Patches\SpawnRagdollEvent.cs" />
     <Compile Include="Patches\SyncDataEvent.cs" />
     <Compile Include="Patches\DecontaminationEvent.cs" />
     <Compile Include="Patches\DoorInteractionEvent.cs" />
     <Compile Include="Patches\DropItemEvent.cs" />
     <Compile Include="Patches\FemurEnterEvent.cs" />
-    <Compile Include="Patches\Generator079Events.cs" />
+    <Compile Include="Patches\GeneratorEvents.cs" />
     <Compile Include="Patches\GhostmodePatch.cs" />
     <Compile Include="Patches\GrenadeThrownEvent.cs" />
-    <Compile Include="Patches\HandcuffEvent.cs" />
+    <Compile Include="Patches\PlayerHandcuffEvents.cs" />
     <Compile Include="Patches\IntercomSpeakingEvent.cs" />
     <Compile Include="Patches\PickupItemEvent.cs" />
     <Compile Include="Patches\PlayerHurtEvent.cs" />
     <Compile Include="Patches\PlayerJoinEvent.cs" />
     <Compile Include="Patches\PlayerLeaveEvent.cs" />
     <Compile Include="Patches\PlayerSpawnEvent.cs" />
-    <Compile Include="Patches\PreAuthEventPatch.cs" />
-    <Compile Include="Patches\GameConsoleQueryOverride.cs" />
-    <Compile Include="Patches\RACommandPatch.cs" />
-    <Compile Include="Patches\ReloadEvent.cs" />
-    <Compile Include="Patches\RespawnEvent.cs" />
-    <Compile Include="Patches\RoundEndPatch.cs" />
+    <Compile Include="Patches\PreAuthEvent.cs" />
+    <Compile Include="Patches\ConsoleCommandEvent.cs" />
+    <Compile Include="Patches\RemoteAdminCommandEvent.cs" />
+    <Compile Include="Patches\PlayerReloadEvent.cs" />
+    <Compile Include="Patches\TeamRespawnEvent.cs" />
+    <Compile Include="Patches\RoundEndEvent.cs" />
     <Compile Include="Patches\RoundRestartEvent.cs" />
-    <Compile Include="Patches\RoundStartPatch.cs" />
-    <Compile Include="Patches\Scp096ProcessLookingOverride.cs" />
+    <Compile Include="Patches\RoundStartEvent.cs" />
+    <Compile Include="Patches\Scp096Events.cs" />
     <Compile Include="Patches\Scp106ContainEvent.cs" />
-    <Compile Include="Patches\Scp106PlayerScriptOverride.cs" />
-    <Compile Include="Patches\Scp106PocketDimOverride.cs" />
-    <Compile Include="Patches\Scp106PocketDimTeleportOverride.cs" />
-    <Compile Include="Patches\Scp173Override.cs" />
+    <Compile Include="Patches\Scp106TeleportEvent.cs" />
+    <Compile Include="Patches\PocketDimensionEvents.cs" />
+    <Compile Include="Patches\PocketDimEscapedEvent.cs" />
+    <Compile Include="Patches\Scp173Patch.cs" />
     <Compile Include="Patches\Scp914ActivationEvent.cs" />
     <Compile Include="Patches\Scp914KnobChangeEvent.cs" />
-    <Compile Include="Patches\ServerCommandPatch.cs" />
+    <Compile Include="Patches\ServerRemoteAdminCommandEvent.cs" />
     <Compile Include="Patches\ServerNamePatch.cs" />
     <Compile Include="Patches\SetClassEvent.cs" />
     <Compile Include="Patches\SetGroupEvent.cs" />
@@ -228,11 +232,11 @@
     <Compile Include="Patches\SmartClassPickerFix.cs" />
     <Compile Include="Patches\StartItemsEvent.cs" />
     <Compile Include="Patches\TriggerTeslaEvent.cs" />
-    <Compile Include="Patches\UseMedicalEvent.cs" />
+    <Compile Include="Patches\UseMedicalItemEvent.cs" />
     <Compile Include="Patches\WaitingForPlayersEvent.cs" />
-    <Compile Include="Patches\WarheadDetonationPatch.cs" />
-    <Compile Include="Patches\WarheadKeycardEvent.cs" />
-    <Compile Include="Patches\WeaponManagerOverride.cs" />
+    <Compile Include="Patches\WarheadEvents.cs" />
+    <Compile Include="Patches\WarheadKeycardAccessEvent.cs" />
+    <Compile Include="Patches\ShootEvents.cs" />
     <Compile Include="Plugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReloadCommandHandler.cs" />
