+++ EXILED_Events/Patches/AntiFlyPatch.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 namespace EXILED.Patches
 {
 	[HarmonyPatch(typeof(PlyMovementSync), nameof(PlyMovementSync.AntiFly), typeof(Vector3))]
-	public class AntiAntiFly
+	public class AntiFlyPatch
 	{
 		public static bool Prefix(PlyMovementSync __instance, ref Vector3 pos) => EventPlugin.AntiFlyPatchDisable;
 	}
