+++ EXILED_Events/Patches/RemoteAdminCommandEvent.cs
@@ -5,7 +5,7 @@ using System;
 namespace EXILED.Patches
 {
 	[HarmonyPatch(typeof(CommandProcessor), nameof(CommandProcessor.ProcessQuery), typeof(string), typeof(CommandSender))]
-	public class OnCommandPatch
+	public class RemoteAdminCommandEvent
 	{
 		public static bool Prefix(ref string q, ref CommandSender sender)
 		{
