+++ EXILED_Events/Patches/SpawnRagdollEvent.cs
@@ -1,0 +1,18 @@
+﻿using Harmony;
+using UnityEngine;
+
+namespace EXILED.Patches
+{
+	[HarmonyPatch(typeof(RagdollManager), nameof(RagdollManager.SpawnRagdoll))]
+	public class SpawnRagdollEvent
+	{
+		public static bool Prefix(RagdollManager __instance, ref Vector3 pos, ref Quaternion rot, ref int classId, ref PlayerStats.HitInfo ragdollInfo, ref bool allowRecall, ref string ownerID, ref string ownerNick, ref int playerId)
+		{
+			bool allow = true;
+
+			Events.InvokeSpawnRagdoll(__instance.gameObject, ref pos, ref rot, ref classId, ref ragdollInfo, ref allowRecall, ref ownerID, ref ownerNick, ref playerId, ref allow);
+
+			return allow;
+		}
+	}
+}
